Moved types to where it belongs
